feat: Phase B——Session/Mutation 数据完整性（B1-B4 全部）
按 TPI_PRODUCTION_HARDENING_ROADMAP Phase B 完成四项 invariant。

B1 Tool-call 生命周期闭合（finalize_unexecuted_calls）：
- invariant: Every model-emitted tool_call eventually has exactly one
  terminal result
- 新增 finalize_unexecuted_calls(session)：扫描事件流中 ToolRequested 无
  ToolCompleted 的 call，best-effort 合成 ToolCompleted(Interrupted)
  （复用 recovery::interrupted_outcome/classify_effect 形态）
- run_inner 的 execute().await 错误路径统一接入（inspect_err）——wave
  循环内 session 写失败/工具 panic 不再留下悬空 tool call；同进程
  refresh_from_log 不再把非法序列发给 provider
- classify_effect 从 pub(crate) 改 pub（跨 crate 复用）
- contract test：finalize 补齐终态 + 幂等（恰好一次）

B2 Undo/Redo CAS：
- invariant: Undo never overwrites content that was not produced by the
  target mutation
- undo_mutation/redo_mutation/undo_all/redo_all 全部改为 CAS 三态：
  current==after→Applied / current==before→AlreadyDone / 其他→Conflict
  不写文件（原子性：任一文件 Conflict 整批拒绝，禁止 C→A）
- 返回 Vec<(path, CasVerdict)> 三态；CLI 输出 Applied/AlreadyDone/Conflict
- contract tests：Applied/AlreadyDone/Conflict/原子性/redo 双向

B3 Journal Tainted：
- load_journal 返回 JournalState{mutations, integrity: Clean|Tainted,
  corrupt_lines}——损坏行不再静默丢弃
- assert_can_mutate：Tainted 拒绝 destructive undo/redo（--force 放行）
- CLI undo/redo 加 --force；崩溃恢复提示显示 Tainted 状态
- contract tests：损坏行→Tainted 拒绝 / Clean 放行

B4 Windows PathIdentity：
- 统一 path_identity()（Windows 大小写折叠）——调度锁（含严格模式）、
  snapshot store key、journal 聚合共用
- 修复严格模式锁不折叠（ISSUE-004 竞态残留）
- contract tests：严格模式 casefold / identity 统一 / journal case 聚合

验证：全 workspace 50+ 套件全绿（agent 78 / capabilities 185 / session
42）；fmt/clippy 干净（仅既有 can-be-collapsed 风格警告）。

+++ crates/tpi-capabilities/src/tool/web.rs
@@ -678,7 +678,7 @@ async fn search_exa(
     let mut args = serde_json::json!({
         "query": query,
         "type": "auto",
-        "numResults": count as u32,
+        "numResults": count,
         "livecrawl": "fallback",
         "contextMaxCharacters": EXA_CONTEXT_MAX_CHARS,
     });
